pc kernel: split mpstartap() and squidboy into separate file... stuff for amd64

+++ sys/src/9/pc/dma.c
@@ -195,7 +195,7 @@ dmasetup(int chan, void *va, long len, int flags)
 	 *  if this isn't kernel memory or crossing 64k boundary or above 16 meg
 	 *  use the bounce buffer.
 	 */
-	if((ulong)va < KZERO 
+	if((uintptr)va < KZERO 
 	|| ((pa=PADDR(va))&0xFFFF0000) != ((pa+len)&0xFFFF0000)
 	|| pa >= 16*MB){
 		if(xp->bva == nil)
